moved getDatafForViewport to API
